+++ src/azure/secret_getter.go
@@ -16,6 +16,7 @@ type SecretGetter struct {
 	client *azsecrets.Client
 }
 
+// TODO - refactor out of the Azure folder and make something domain-specific
 func NewSecretGetter() (SecretGetter, error) {
 
 	// Create a credential using the NewDefaultAzureCredential type.
