+++ src/usecases/blob.go
@@ -1,4 +1,4 @@
-package main
+package usecases
 
 // The BlobHandler interface is about interacting with file data,
 // e.g. in Azure Blob Storage or a local filesystem
