+++ src/usecases/sender.go
@@ -1,4 +1,4 @@
-package main
+package usecases
 
 // The MessageSender interface is about delivering data to external services.
 // Currently, we send messages to ReportStream or to a local-only mock service for testing.
